+++ datafusion/catalog/src/memory/mod.rs
@@ -17,6 +17,8 @@
 
 pub(crate) mod catalog;
 pub(crate) mod schema;
+pub(crate) mod table;
 
 pub use catalog::*;
 pub use schema::*;
+pub use table::*;
